move ModuleObj to new module.py file
