+++ bifacial_radiance/__init__.py
@@ -1,5 +1,6 @@
-from bifacial_radiance.main import AnalysisObj, GroundObj, MetObj, RadianceObj, SceneObj, ModuleObj
+from bifacial_radiance.main import AnalysisObj, GroundObj, MetObj, RadianceObj, SceneObj
 #from bifacial_radiance.readepw import readepw
+from bifacial_radiance.module import ModuleObj
 from bifacial_radiance import load
 from bifacial_radiance import modelchain
 from bifacial_radiance.gui import gui
